# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Online Bookstore API",
        "description": "Online Bookstore Management System",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/cart/order": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Order Add",
                "operationId": "0be414b98aa0894d9061ed7cbc073901",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "book_id": {
                                                "type": "integer"
                                            },
                                            "qty": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "book_id": 1,
                                    "qty": 2
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "order_id": {
                                                    "type": "string"
                                                },
                                                "customer": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "number",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "test"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "test@gmail.com"
                                                        },
                                                        "role": {
                                                            "type": "string",
                                                            "example": "Customer"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "item": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "book_id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "qty": {
                                                                "type": "integer",
                                                                "example": 2
                                                            },
                                                            "sub_total": {
                                                                "type": "boolean",
                                                                "example": 240000
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 240000
                                                },
                                                "payment_method": {
                                                    "type": "string",
                                                    "example": null
                                                },
                                                "address": {
                                                    "type": "text",
                                                    "example": null
                                                },
                                                "delivery_service": {
                                                    "type": "string",
                                                    "example": null
                                                },
                                                "is_paid": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "payment_proof": {
                                                    "type": "string",
                                                    "example": null
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "Pending"
                                                },
                                                "is_completed": {
                                                    "type": "boolean",
                                                    "example": false
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}